requirements.txt
numpy
matplotlib
pandas
pyyaml
mlflow
fastapi
uvicorn